requirements.txt
SQLAlchemy
mysqlclient
ddt
google-cloud-bigquery